style: apply clippy lint fixes across non-TUI crates
Mechanized code-quality improvements driven by clippy:
- struct update syntax (`..Default::default()`) instead of mutable-then-assign
- float equality via epsilon comparison instead of `==`
- `is_some_and`/`is_none_or` instead of `.map().unwrap_or()`
- explicit enum variants instead of wildcard matches
- remove unnecessary `.clone()` and leading-underscore bindings
- inline `use` imports at first use; move helper structs before use
- backtick formatting in doc comments; remove redundant `INLINE_CODE` flag
- replace `vec![x]` literals with slice references where appropriate

+++ crates/y-tools/src/builtin/file_read.rs
@@ -968,7 +968,7 @@ mod tests {
         let dir = tempfile::tempdir().unwrap();
         let file_path = dir.path().join("too_many_tokens.txt");
         let mut f = std::fs::File::create(&file_path).unwrap();
-        writeln!(f, "{}", "x".repeat((MAX_OUTPUT_TOKENS as usize + 1) * 4)).unwrap();
+        writeln!(f, "{}", "x".repeat((MAX_OUTPUT_TOKENS + 1) * 4)).unwrap();
 
         let tool = FileReadTool::new();
         let input = make_input(serde_json::json!({
